requirements.txt
torch
torchvision>=0.14
numpy
Pillow
scikit-image
pyyaml
tqdm
streamlit
matplotlib
opencv-python
tensorboard
reportlab
cryptography

